Django 4 Incompatibility fixed
Fixed bug in which DISTINCT clauses were forcibly removed from supplied  querysets

I was supplyin a queryset which was already distcint! And moreover I am using .distinct(field) in order to make a complex inetraction between url filters and Window functions work. Works a dream in fact except that url_filter was clobberig the necessary DISTINCT ON clause in my query.

THis is where it happened and it was because it paid no heed to the existing status of the querie's distinct() setting. It neeed only apply distinct() if the caller has anot already applied on (not least as called knows what fields they need to distinguish on.

Fixed a typo in the Makefile

Simple spelling mistake. It's quiet not quite. Well a comix might say it's quite quiet ;-).